Return non-zero exit status in case of failure
When compilation fails or qb encounters other errors, exit the program
with the exit code of 1 instead of just doing `return` from main which
seems to result in the exit code 0.

This will enable doing things like `qb && ./program` in order to compile
the program and then run it if but only if compilation was successful.